initial react architecture redesign
