+++ client/App.jsx
@@ -1,9 +1,10 @@
 import React from "react";
 import { BrowserRouter as Switch, Route } from "react-router-dom";
-import Signup from "./components/Signup";
-import Login from "./components/Login";
+import Signup from "./routes/Signup";
+import Login from "./routes/Login";
+import ChatContainer from "./routes/ChatContainer";
 import LBar from "./components/LBar.jsx";
-import Basepage from "./components/Basepage.jsx";
+import Basepage from "./routes/Basepage.jsx";
 // import "./stylesheets/styles.css";
 
 function App() {
@@ -11,7 +12,7 @@ function App() {
     <div className="app">
       <Switch>
         <Route exact path="/" component={Basepage} />
-        <Route path="/window" component={LBar} />
+        <Route path="/window" component={ChatContainer} />
         <Route path="/login" component={Login} />
         <Route path="/signup" component={Signup} />
       </Switch>
